docs: fronted page factory

+++ bc_obps/reporting/api/report_attachments.py
@@ -2,7 +2,7 @@ from typing import List, Literal, Tuple
 from common.api.utils.current_user_utils import get_current_user_guid
 from common.permissions import authorize
 from django.db import transaction
-from django.http import HttpRequest, StreamingHttpResponse
+from django.http import HttpRequest
 from ninja import File, Form, UploadedFile
 from registration.decorators import handle_http_errors
 from reporting.constants import EMISSIONS_REPORT_TAGS
@@ -15,7 +15,7 @@ from .router import router
 
 @router.post(
     "report-version/{report_version_id}/attachments",
-    response={200: int, custom_codes_4xx: Message},
+    response={200: List[ReportAttachmentOut], custom_codes_4xx: Message},
     tags=EMISSIONS_REPORT_TAGS,
     description="""Saves the reporting attachments, passed as text in the payload.""",
     auth=authorize("approved_industry_user"),
@@ -27,25 +27,34 @@ def save_report_attachments(
     report_version_id: int,
     file_types: Form[List[str]],
     files: List[UploadedFile] = File(...),
-) -> Literal[200]:
+) -> Tuple[Literal[200], List[ReportAttachmentOut]]:
+
+    if len(file_types) != len(files):
+        raise IndexError("file_types and files parameters must be of the same length")
 
     for index, file_type in enumerate(file_types):
 
-        report_attachment = ReportAttachment.objects.get(report_version_id=report_version_id, attachment_type=file_type)
+        report_attachment: ReportAttachment | None = ReportAttachment.objects.filter(
+            report_version_id=report_version_id, attachment_type=file_type
+        ).first()
 
-        # Delete file from storage then from db
-        report_attachment.attachment.delete()
-        report_attachment.delete()
+        # Delete file from storage then from the db
+        if report_attachment:
+            report_attachment.attachment.delete(save=False)
+            report_attachment.delete()
+
+        file = files[index]
 
         attachment = ReportAttachment(
             report_version_id=report_version_id,
-            attachment=files[index],
+            attachment=file,
             attachment_type=file_type,
+            attachment_name=file.name or "attachment",
         )
         attachment.save()
         attachment.set_create_or_update(get_current_user_guid(request))
 
-    return 200
+    return load_report_attachments(request, report_version_id)
 
 
 @router.get(
@@ -59,15 +68,5 @@ def load_report_attachments(
     request: HttpRequest,
     report_version_id: int,
 ) -> Tuple[Literal[200], List[ReportAttachmentOut]]:
-    return ReportAttachment.objects.filter(report_version_id=report_version_id)
-
-
-@router.get(
-    "report-version/{report_version_id}/attachments/{file_id}",
-)
-def download_report_attachment_file(request: HttpRequest, report_version_id: int, file_id: int):
-
-    file = ReportAttachment.objects.get(id=file_id, report_version_id=report_version_id)
-
-    response = StreamingHttpResponse(streaming_content=file)
-    response['Content-Disposition'] = f'attachment; filename={file.name}'
+    attachments_data = ReportAttachment.objects.filter(report_version_id=report_version_id).all()
+    return 200, attachments_data  # type: ignore
